requirements.txt
google-cloud-aiplatform
google-cloud-storage
pandas==2.2.2
matplotlib
folium
plotly
kaleido==0.2.1
numpy
vertexai